p2 good results

+++ deep-reinforcement-learning/p2_continuous-control/agents/replay_buffers.py
@@ -182,7 +182,7 @@ class PrioritizedReplayBuffer(object):  # stored as ( s, a, r, s_ ) in SumTree
             b_idx[i] = idx
         
         states = torch.from_numpy(np.vstack([e.state for e in b_memory if e is not None])).float().to(self.device)
-        actions = torch.from_numpy(np.vstack([e.action for e in b_memory if e is not None])).long().to(self.device)
+        actions = torch.from_numpy(np.vstack([e.action for e in b_memory if e is not None])).float().to(self.device)
         rewards = torch.from_numpy(np.vstack([e.reward for e in b_memory if e is not None])).float().to(self.device)
         next_states = torch.from_numpy(np.vstack([e.next_state for e in b_memory if e is not None])).float().to(self.device)
         dones = torch.from_numpy(np.vstack([e.done for e in b_memory if e is not None]).astype(np.uint8)).float().to(self.device)
